pyproject.toml
[project]
name = "nbac"
version = "0.1.0"
description = "A simple Python package for Naive Bayes archetype classification (NBAC)."
readme = "README.md"
requires-python = ">=3.8"
dependencies = [
  "cloudflare",
  "psycopg2-binary",
  "python-dotenv", # optional
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
